% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.sdad.health.outcomes-package.R
\docType{package}
\name{dc.sdad.health.outcomes-package}
\alias{dc.sdad.health.outcomes}
\alias{dc.sdad.health.outcomes-package}
\title{dc.sdad.health.outcomes: Social Data Commons: Health, Behavior, Diet: Health Outcomes}
\description{
Allows user to easily get Health Outcomes Derived from the Behavioral Risk Factor Surveillance System data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Eric Oh \email{eo9tv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
